Add new `metrics` package compatible with API v0.8
The old `metric` package (not the singular) is based on the old API
v0.5. Since the package name didn't match the API naming scheme, we add
a new package with the right name to keep backwards compatibility.

These classes are based on the microgrid API client v0.18:
https://github.com/frequenz-floss/frequenz-client-microgrid-python/tree/v0.18.0/src/frequenz/client/microgrid/metrics

Only some minor changes were applied to make it closer to the API names,
for example, `MetricSample.sampled_at` was renamed to `.sample_time` to
match the new name.

Also the `MetricConnection` and `MetricConnectionCategory` classes were
added.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>